refactor move exampleBrickConfigs -> bricks/ directory
