+++ src/analysis/analysisVisitors/brickTypeAnalysis.test.ts
@@ -19,7 +19,7 @@ import BrickTypeAnalysis from "@/analysis/analysisVisitors/brickTypeAnalysis";
 import brickRegistry from "@/bricks/registry";
 import CommentEffect from "@/bricks/effects/comment";
 import { sidebarPanelFormStateFactory } from "@/testUtils/factories/pageEditorFactories";
-import { createNewConfiguredBrick } from "@/pageEditor/exampleBrickConfigs";
+import { createNewConfiguredBrick } from "@/bricks/exampleBrickConfigs";
 import { DocumentRenderer } from "@/bricks/renderers/document";
 import CancelEffect from "@/bricks/effects/CancelEffect";
 import { ALERT_EFFECT_ID, AlertEffect } from "@/bricks/effects/alert";
