+++ src/bricks/exampleBrickConfigs.ts
