+++ src/pageEditor/fields/FormModalOptions.test.tsx
@@ -19,7 +19,7 @@ import React from "react";
 import { render } from "@/pageEditor/testHelpers";
 import FormModalOptions from "@/pageEditor/fields/FormModalOptions";
 import registerDefaultWidgets from "@/components/fields/schemaFields/widgets/registerDefaultWidgets";
-import { createNewConfiguredBrick } from "@/pageEditor/exampleBrickConfigs";
+import { createNewConfiguredBrick } from "@/bricks/exampleBrickConfigs";
 import { FormTransformer } from "@/bricks/transformers/ephemeralForm/formTransformer";
 import { screen } from "@testing-library/react";
 import { actions as editorActions } from "@/pageEditor/store/editor/editorSlice";
